8085: mini boot ROM fixes
With these done it's possible to get the 8085 board, CF adapter and the
512K ROM/RAM board up and running together (tested on real hardware). The
only supported serial right now is a 16550A at 0xC0. In theory the SIO and
ACIA are both usable.

The 8085 board has a design error that makes interrupts break but this can
be fixed with a small board mod (See the Fuzix docs for rc2014-8085).

Note that the 8085 board provides no way to generate M1 cycles so you cannot
fully use the Z80PIO or Z80CTC card with it.