simple80: workaround for contention via resistors
Probably the data bus resistors between the ROM and RAM should be 470Ohm or
more but a workaround is to write zero to the RAM, map the ROM and then
read it, then restore the byte in RAM.

Ick!

At some point I need to build a second board with bigger resistors.