rom: clean up so we can build it with zmac on CP/M
This gets us close to self hosting.

ZMAC ROM.ASM
LINK ROM[L0]=ROM

So now all that is needed is an RC2014 compatible EPROM burner and something
to trim it to the right size. Well Ok you can do it with

DDT ROM.COM
G0
SAVE 32 ROM.BIN

but that's ugly 8)

For CP/M we also need a tool to put the CPMLDR on the right blocks