68000: Update the ROM to v 0.03
Bring the system up to the point PPIDE works and we also support enough
ROM to RAM revectoring of exceptions for Fuzix to run. The boot loader
environment is also now a little different and sufficient to allow console
calls and raw disk block I/O calls to be made by any boot loader.

Boot loading is also no longer automatic - switch to the needed volume and
type BOOT at the prompt.